Add structure and initial implementation
