+++ .gitignore
@@ -1,5 +1,6 @@
 # Prerequisites
 *.d
+build/
 
 # Compiled Object files
 *.slo
