+++ Makefile
@@ -27,8 +27,11 @@ clean:
 all: init
 	$(CMAKE) --build $(BUILD_DIR) --target all -- -j$(JOBS)
 
-build-test:
-	$(CMAKE) --build $(BUILD_DIR) --target tests.tsk -- -j$(JOBS)
+
+$(BUILD_DIR)/test/unit/tests.tsk:
+	$(CMAKE) --build $(BUILD_DIR) --target tests -- -j$(JOBS)
+
+build-test: init $(BUILD_DIR)/test/unit/tests.tsk
 
 exec-test:
 	@cd $(BUILD_DIR) && \
